refactor: remove hostReviews and eventReviews from host profile fetch

+++ src/app/modules/v1/host/host.validation.ts
@@ -1,7 +1,7 @@
 import z from "zod";
 
 export const reviewHostSchema = z.object({
-  rating: z.number().int().min(1).max(5, "Rating must be between 1 and 5"),
+  rating: z.number().min(1).max(5, "Rating must be between 1 and 5"),
   comment: z
     .string()
     .min(10, "Comment must be at least 10 characters")
